Move prototype musig2 implementation to commonTest
